Fixes for graph view and gutter (folding)
Phew. The gutter folding thing was fun (thanks @Silver for all your help there)
Fixed the colours in the graph view and a couple of other minor issues like the header wrapping (thanks @ryanjmurphy). Though in that case, the header overlaps content when it's more than four lines long, so still room for improvement