Colour Revamp!
I revamped all the colours with my learnings from my Base2Tone theme.
Right near the top of the CSS file, there's a hue and saturation value you can play with that will recolour the whole interface.
It's simple to change the lightness values (L0-L3) and/or text-colour for light and dark themes separately, also.

Finally, if you want to use this theme with AndyV2 (or just don't like the look of the in-body-style header and want the Obsidian-style headers back) just delete everything below the comment: `/* !!! Remove from here down if you want to use with Andy Matuschak mode V2 !!! */`