Removed changes to heading and preview to line up with source
The source editor doesn't line up with the heading anymore with this change, depending on if there's a gutter. If there's no gutter it's fine.
You get a gutter if you enable folding and/or line numbers, and depending on which you enable, the gutter's size is different, so there's not really much point in me trying to compensate for it when it could be different for everyone.