Made "inline header" only apply to markdown documents
The inline header code was messing with the layout of the graph view, so I made them only apply to markdown documents, and more "traditional" headers will appear for graph view, image files, etc.
I also removed the code that added padding to the bottom of the markdown panes, as that was added to obsidian itself in a previous update.